[BE] Add: 1:1 채팅 구현 완료

+++ backend/src/main/java/com/ssafy/honjaya/db/repository/ChatRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ssafy.honjaya.db.entity.Chat;
 
-public interface ChatRepository extends JpaRepository<Chat, Integer> {
-	List<Chat> findAllByChatroom_ChatroomNo(int chatroomNo);
+public interface ChatRepository extends JpaRepository<Chat, Long> {
+	List<Chat> findAllByChatroom_ChatroomNo(long chatroomNo);
 }
